+++ src/main/java/com/example/kinalapp/repositry/DetalleVentaRepository.java
@@ -1,11 +1,11 @@
 package com.example.kinalapp.repositry;
 
-import com.example.kinalapp.Service.DetalleVenta;
+import com.example.kinalapp.entity.DetalleVenta;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
 public interface DetalleVentaRepository extends JpaRepository <DetalleVenta, String> {
 
-    List<com.example.kinalapp.entity.DetalleVenta> findByEstado(int estado);
+    List<com.example.kinalapp.entity.DetalleVenta> findByEstado(Long estado);
 }
